Dvr3

+++ include/vapor/DVRenderer.h
@@ -20,7 +20,7 @@ class RENDER_API DVRenderer : public RayCaster {
 
   protected:
     int _load3rdPassShaders();
-    void _3rdPassSpecialHandling(bool fast, long castMode);
+    void _3rdPassSpecialHandling(bool fast, int castMode);
 };
 
 }; // namespace VAPoR
