+++ include/vapor/IsoSurfaceRenderer.h
@@ -20,7 +20,7 @@ class RENDER_API IsoSurfaceRenderer : public RayCaster {
 
   protected:
     int _load3rdPassShaders();
-    void _3rdPassSpecialHandling(bool fast, long castMode);
+    void _3rdPassSpecialHandling(bool fast, int castMode);
 };
 
 }; // namespace VAPoR
